tast-tests: syzcorpus: allow for test debugging
The tast test has been showing different results while running in
postsubmit compared to local tests.

To further diagnoze the issue, reset the DUT before starting the
test; and set windowSize to 1. This is a temporary measure
and should be removed in the future.

BUG=b:192782710
TEST=tast -verbose run localhost:2200 syzcorpus.Kvmtest

Change-Id: I07084346753670762c2dddae71e21f76526e86f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3005093
Commit-Queue: Zubin Mithra <zsm@chromium.org>
Tested-by: Zubin Mithra <zsm@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>