tast-tests: Initialize UIAutomator in arcBooted
Many ARC Tast tests rely on UIAutomator. However, UIAutomater is
initialized in each test, resulting in additional execution time and
flakiness.
This CL initializes UIAutomator in arcBooted and updates all the tests
relying on UIAutomator to use that. The list of the tests that had to be
updated was obtained by:

    rg -l -U 'arcBooted(?s).*NewUIDevice'

BUG=b:190014347
TEST=tast run hatch-arc-r-DUT arc.InputCompat.vm arc.KeyCharacterMap.vm

Change-Id: Iaa9a3c72fc19d61c331dee53905823111f4f5c0b