[CrOS WebAPKs] Add Android app to test sharing to WebAPKs
This Android app just shows a button which shares some text to any
installed WebAPK that accepts text shares.

This will be used in a test where we install a WebAPK and assert that an
Android app is able to share content through the WebAPK to a Web App.

BUG=chromium:1226730
TEST=emerge-$BOARD tast-local-apks-cros (Then installed the APK
  manually and verified the behavior)

Change-Id: Ie47439548b89e1ac4ef4cce7b11612b28c7d34c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3010632
Tested-by: Tim Sergeant <tsergeant@chromium.org>
Auto-Submit: Tim Sergeant <tsergeant@chromium.org>
Reviewed-by: Yuichiro Hanada <yhanada@chromium.org>
Commit-Queue: Tim Sergeant <tsergeant@chromium.org>