[CrOS WebAPKs] Add test PWA and pregenerated WebAPK for that PWA
This adds a basic PWA, based off the PWA used for apps.Sharesheet, to
the appsplatform data folder. This will allow us to write a test which
shares data from Android to the PWA (again, similar to apps.Sharesheet).

This also adds a pregenerated WebAPK representing the test PWA. Under
ordinary circumstances this WebAPK would be automatically generated
and installed when the PWA is installed. For the moment, we'll instead
install a pregenerated WebAPK while we determine how to make the real
installation flow stable in a Tast environment.

BUG=chromium:1226730
TEST=Manual (copied PWA files to /tmp on DUT, `python -m http.server`,
  installed PWA and WebAPK, verified behavior)

Change-Id: I9e3f42ec4207fb6a19c9762ce453f28cc9e336f0