Graphics.IGT: Stop tlsdated before the tests
Few IGT tests like kms_flip does tests in combination with suspend.
For this, waking up is achieved using the RTC wake-up alarm.

If any other service is holding /dev/rtc, the IGT test fails to take the lock and set a wake up alarm. Hence, it is required to stop the service which is currently holding the RTC lock before running the IGT test.

In case of chrome, tlsdated holds rtc lock. This patch stops the service before IGT tests are started.

BUG=b:193034528
TEST=cros_sdk tast run VOLTEER graphics.IGT.kms_flip

Change-Id: Iea6a9eec0b8d4f459a6e6ea043716ea90a2f23b3