wifi: Port network_WiFi_ReconnectInDarkResume to Tast
Port the old lucid_sleep test to Tast.

Drop the dark resume count check as it highly depends on the internal
implementation and the AP configuration time. Instead, we just expect
the DUT to reconnect in a reasonable time. e.g. 1 min

note: currently disconnect_before_suspend_* variation will fail, which
might be related to b/188743774.

BUG=b:189407934
TEST=`tast run $DUT wifi.ReconnectInDarkResume.*` on volteer

Change-Id: I4e6093415e7cb33a1b57838c1c57900055672a56