WIP: Create Tast tests for powerd plugin on fwupd
This CL will create tests that run commands or processes to test the
behavior and efficacy of a powerd plugin for fwupd.

THe first test, written in fwupd_powerd_startup.go, runs the command
"fwupdmgr get-plugins" and searches for the key word "powerd" in order
to test whether it shows up on the list on enabled plugins.

Currently, the test mentioned below fails with the following error:
"Failed to run tests: no tests matched by pattern(s)
[fwupd_powerd_startup.go]"

BUG=b:193158345
TEST= tast run [user@]host[:port] fwupd_powerd_startup.go

Change-Id: I7490d457bc29132e940ab354823689a6351e5310