fwupd: Create Tast tests for powerd plugin on fwupd
This CL creates a test that runs a command or process to gauge the
behavior and efficacy of a powerd plugin for fwupd.

This test, in particular, runs the command "fwupdmgr get-plugins"
and searches for the key phrase "powerd Flags: ? Enabled" (with the appropriate whitespace included)
in order to test whether it shows up on the list on enabled plugins.

The test passes when the command mentioned below is run and it can identify this string
in the output.

BUG=b:193158345
TEST=tast run [user@]host[:port] firmware.FwupdPowerdStartup

Change-Id: I7490d457bc29132e940ab354823689a6351e5310