policy: Wait for ARC to start up before logging in
The ARC container mini instance, started when the login screen is shown,
blocks session_manager and prevents it from responding to D-Bus methods.
Cloud policy initialisation in Chrome relies on being able to contact
session_manager. If it can't complete policy initialisation within 10
seconds, the user is logged out, which is causing tests to fail.

Waiting for ARC to finish starting essentially waits for session_manager
to be free to respond to other requests.

BUG=b:192195817
TEST=`tast run $DUT policy.AllowDeletingBrowserHistory`

Change-Id: Ic839aef2587ac5b39317054547fc0d0c90682a17