Disable failing camera test cases on reven.
Reven runs kernel-5_4, but the vivid kernel driver that a few of the
camera tasts require is broken in that version (see b/189483008).

This changeset skips the failing test cases for now: this will allow us
to remove a hack that we introduced previously that skips the entire
group:camera-libcamera.

BUG=b:192644384
TEST=tast run <DUT> group:camera-libcamera

Change-Id: I44d44ec3dfb7d15d1230f7b94fd14c4690e40c77