tast-tests: power: Add NewSignalWatcher utility.
Wake on WiFi feature depends on dark resume and its tests will need to
verify that dark resumes are [not] triggered on specific cases.

Export the signal names with constants and extract a utility for
spawning the D-Bus signal watcher, so that other tests can share the
logic.

BUG=b:189409165
TEST=tast run $DUT wilco.APIHandlePowerEvent

Change-Id: I8bf91e21d48404d578f3d74f435c870089ca439e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3034705
Tested-by: Yen-lin Lai <yenlinlai@chromium.org>
Commit-Queue: Yen-lin Lai <yenlinlai@chromium.org>
Reviewed-by: Oleh Lamzin <lamzin@google.com>