google meet: wait for the menu expansion animation to complete
When creating new google slide online file, when we click "New"
button in the web page, some DUTs will see the new menu
stuck in the middle of its animation - it failed to expend to
its full size in couple of seconds.
This CL tries to fix this issue by waiting for menu expansion to
complete.

BUG=b:193425635
TEST=tast run <vars> <dut> ui.GoogleMeetCUJ.*

Change-Id: Ibae6119be9187d540b7dd1ce88973e0cc9ce8085
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3048765
Tested-by: Xibin Liu <xibin@google.com>
Commit-Queue: Xibin Liu <xibin@google.com>
Reviewed-by: Jun Sheng <shengjun@chromium.org>