faft: Firmware precondition funcs for test params
Every test that uses the firmware precondition to set the boot mode
needs to specify Data, ServiceDeps, SoftwareDeps, and Vars. Adding a
new required dep there is also painful as every test has to be modified.

Instead add helper functions to provide the vars that the precondition
needs but the test itself doesn't care about.

BUG=b:190532175
TEST=~/trunk/src/platform/tast/fast_build.sh -T

Change-Id: Iceda18f40ddc4c24f7187bfec10d61ddc6e90978