tast-tests/video: Add ImageProcessor tast test
image_processor_test is a binary unit test for the ImageProcessor
module. On systems with a compatible GPU ImageProcessor backend, it will
use the GPU to perform colorspace conversion, scaling and rotation
operations. On other systems, it defaults to using a libYUV-based
ImageProcessor backend.

This commit adds a tast test to run this test binary.

BUG=b:182952329
TEST=tast run ${IP} video.ImageProcessor.*

Depends-On: I3a5bac4c9094cfda4c9bcdace8aba89b487f452f
Change-Id: I14bb5e4569f822ac8f8bc4d91f6690567097193f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3032944
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>
Reviewed-by: Po-Hsien Wang <pwang@chromium.org>
Tested-by: Nathan Hebert <nhebert@chromium.org>
Commit-Queue: Nathan Hebert <nhebert@chromium.org>