tast-tests/video: Add ImageProcessor tast test
image_processor_test is a binary unit test for the ImageProcessor
module. On systems with a compatible GPU ImageProcessor backend, it will
use the GPU to perform colorspace conversion, scaling and rotation
operations. On other systems, it defaults to using a libYUV-based
ImageProcessor backend.

This commit adds a tast test to run this test binary.

BUG=b:182952329
TEST=tast run ${IP} video.ImageProcessor

Change-Id: I14bb5e4569f822ac8f8bc4d91f6690567097193f