tast: deflake crostini.VerifyAppWayland
This test uses the container's wayland_demo app to display a
solid color rect. The wayland_demo app can crash if there isn't
enough memory for the shared memory buffer.

We limit the size of the displayed rectangle, and modify the
screenshot checking code to limit its attention to the correct
portion of the screen.

BUG=chromium:1175102
TEST=tast run crostini.VerifyAppWayland.buster_stable

Change-Id: Ia368159598d1a87f2032abc4697b27213126d82f