wificell/pcap: clean up remote files on download
Packet capture devices have limited /tmp storage, and depending on the
environment WiFi tests are run in, tests can collect rather large pcap
files. Over the course of a single Tast run, an unbounded number of
tests can run, and so we potentially accumulate a large number of these
files and run out of space.

Let's be a little more aggressive on cleaning up after ourselves.

Note that these are typically stored in a /tmp/tast-test/ directory that
gets cleaned up each time a router object is initialized, but this
doesn't happen often, especially when we use Fixtures for the test
initialization.

BUG=b:187369955
TEST=wifi.*

Change-Id: I5871884a4ceb2fb01f8dbe030d5d68e0cbe9fa33
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3027682
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Yen-lin Lai <yenlinlai@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>