tast-tests: Improve error handling in arc.WebAPK activity launching
Previously, errors in the middle of the launchTestApp function would not
clean up the intermediate state of the activity. This CL removes the
launchTestApp function and moves the code back into the test function,
which allows us to defer the clean up tasks as we go.  This seems
cleaner than trying to handle all of the different cases in the helper
function.

BUG=chromium:1226730
TEST=tast run $DUT appsplatform.WebAPK.vm

Change-Id: I4e5ea38f0db0fb238654d8010c3d75d3f95cf315
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3032826
Tested-by: Tim Sergeant <tsergeant@chromium.org>
Commit-Queue: Tim Sergeant <tsergeant@chromium.org>
Reviewed-by: Judy Wu <jinrongwu@google.com>