faft: Use a smarter waitUnreachable
When we wait for the DUT to be unreachable, check the ec power state and
stop when we reach S5 or G3. If that doesn't happen quickly, fall back
to DUT.WaitUnreachable.

BUG=b:174846911
TEST=local kindred in normal mode w/ valid usb
tast run $DUT firmware.Fixture*
Ran 4 remote test(s) in 4m46.769s
This function was faster, but something else was slower in this run.

Change-Id: I969042bd5ac172a63df18163874d06ebcd1fdc0f