faft: Don't wait for poweroff command
The poweroff command will always time out, so use a goroutine to wait
for the final result in the background. For some reason sometimes Start
times out, so set a context timeout on it also.

BUG=b:174846911
TEST=local kindred in normal mode w/ valid usb
tast run $DUT firmware.Fixture*
Ran 4 remote test(s) in 5m52.817s
No noticible change from the last run, but sometimes this seems to save
2 seconds in the logs.

Change-Id: I99544185810fe3ee3311b2e3653dfc6027720e59