audio: Fix WaitForDevice function
The CL:2992580 changed the behavior of WaitForDevice. The new function
can not handle the situation when the streamType is
InputStream|OutputStream since the variable 'active' is moved from the
outside to the inside of for-loop. This change can fix this issue.

BUG=b:194034171
TEST=tast run {IP} audio.CrasPerf.playback

Change-Id: I695b42ab45679fb301f98e793717a9e5ffcacb1c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3038225
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-by: Jun Sheng <shengjun@chromium.org>
Commit-Queue: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Tested-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
(cherry picked from commit b89c60c293291a8edd67daec10469bf3dcfb8004)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3051950
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>