tast-tests: Change SearchInHelpApp query to helpa
Why: "helpa" has a prefix match for "help" and prefix matches are more
lenient than fuzzy matches. This means the test will be more resilient
to future changes in the LSS scoring logic.

Also change the expectedResultFinder to look for "Help" in the name.
The previous logic was looking for "help" (case sensitive) in the title
and body snippets. The category is always present, but it is "Help",
so it didn't match.

BUG=chromium:1221017
TEST=Ran the apps.SearchInHelpApp* tests

Change-Id: I6f1705061755cd903f061f1123b9da5f9303d2ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3040761
Reviewed-by: Jun Sheng <shengjun@chromium.org>
Tested-by: Zufeng Wang <zufeng@google.com>
Commit-Queue: Zufeng Wang <zufeng@google.com>