fwupd: Tast test for Absence of Powerd
This CL creates a new Tast test that first disables the power daemon if
it is running. It then runs the command "fwupdmgr get-plugins --json" to get a
list of plugins that were compiled for fwupd. Afterwards, it checks for
the string "Name: powerd, Flags: Disabled" (with the appropriate whitespace),
which is outputted when the plugin is disabled.

The power daemon is the started up again after the output of "fwupdmgr
get-plugins --json" has been searched.

If it passes, then powerd was appropriately disabled
upon compilation since the daemon was absent.

BUG=b:194181934
TEST=tast run [user@]host[:port] firmware.FwupdPowerdDisabled

Change-Id: Id4b9187987a4e9f3b6837123d20c26db6e7748d2