tast-tests: Fix flakiness in appsplatform.WebAPK.vm
This test is about ~50% flaky on ARC VM due to the UI Automator seeming
to prevent the WebAPK launch from going ahead on the ARC side. Closing
the UI Automator server as soon as we're done with it seems to address
the flakiness locally -- I haven't tried to dig in to the root cause.

BUG=chromium:1226730
TEST=tast run $DUT appsplatform.WebAPK.vm

Change-Id: Ic0a7c99b7478fc16bfcddea4365a7df5ee2d0506