tast-tests: Promote ChromeVox hint to critical
The a11y tastboard [1] shows the ChromeVox hint test to be
relatively stable, with a few known failures; these
rare failures seem to be mostly due to loss of SSH connection or
other precondition failures.

As a result, we can confidently promote the tests to critical.

[1] https://tastboard.corp.google.com/test?contact=&testname=a11y&dep=

BUG=chromium:1159107
TEST=tast -verbose run <IP> a11y.ChromevoxHint

Change-Id: Ibfed47d03371024164071c4f9cceb8230ab159c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3045497
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Tested-by: Akihiro Ota <akihiroota@chromium.org>
Reviewed-by: Alexander Hartl <alexanderhartl@google.com>