tast-tests: Temporarily disable RunOCI for noexec
A security change in glibc prevents loading shared libraries from
noexec partitions. Apparently for this test /lib64/librt.so.1
triggers the noexec failure, leading to a different stderr message
and no stdout. But it doesn't seem to be consistent.

For now, this test is getting disabled until we're able to fix
the test on all devices reliably.

Tracking bug is b/194923131.

Must be merged along with chromium:2910526.

BUG=chromium:1182687
TEST=CQ of http://crrev.com/c/2910526

Change-Id: Ifa22506ad3a698d82bda009e77c61f9485a84bda