tast-tests: Temporarily ignore librt in RunOCI
A security change in glibc prevents loading shared libraries from
noexec partitions. Apparently for this test /lib64/librt.so.1
triggers the noexec failure, leading to a different stderr message
and no stdout. But it doesn't seem to be consistent.

For now, we're working around this test by ignoring errors which
have the string /lib64/librt.so.1, at the request of jorgelo@.

Tracking bug is b/194923131.

BUG=chromium:1182687
TEST=CQ of http://crrev.com/c/2910526

Change-Id: Ifa22506ad3a698d82bda009e77c61f9485a84bda
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3046720
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Tested-by: Jordan R Abrahams <ajordanr@google.com>
Commit-Queue: Jordan R Abrahams <ajordanr@google.com>