video: extend PlatformEncoding to software h264 cases
This CL adds the use cases and necesary vars/methods to run
openh264enc, the wrapper around the library used to encode H.264
by e.g. WebRTC. The meat of the CL is the argsOpenh264enc()
method that fills in the command line replicating said WebRTC
configuration.

BUG=b:174103282
TEST=tast run <volteer ip> video.PlatformEncoding.openh264enc*
on volteer (target bitrate is sadly off for reasons I don't
fully comprehend, but the SSIM/PSNR and visual results are good).

Change-Id: Ibbc85cfe5c3456b84707d6330b8f0ed945800554