fwupd: Tast test for powerd battery case handling
This test runs an update command on a fake webcam "device" and checks
what powerd does under 2 different cases. After restarting powerd to
make sure we can run an update command:

In case 1, we simulate a scenario where the computer is not plugged by
forcing it to discharge. This makes the plugin think there is no AC
power, meaning the update should be stopped.

In case 2, we simulate the presence of AC power, so we simply plug the
computer in manually and run the update command, which should pass.

BUG=b:193159451, b:195759844
TEST=tast run [user@]host[:port]
firmware.FwupdPowerdUpdateCheck.[case_name]

Change-Id: I593dfe8d3e702e620c1fc8ec016a28fc292e4f0b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3048242
Tested-by: Daniel Campello <campello@chromium.org>
Reviewed-by: Jun Sheng <shengjun@chromium.org>
Commit-Queue: Daniel Campello <campello@chromium.org>