fwupd: Tast test for powerd battery case handling
This test runs an update command on a fake webcam "device" and checks
what powerd does under 3 different cases.

In case 1, we simulate a scenario where the computer is not plugged by
forcing it to discharge. This makes the plugin think there is no AC
power, meaning the update should be stopped.

In case 2, we simulate the presence of AC power, so we simply plug the
computer in manually and run the update command, which should pass.

In case 3, we throw the "--ignore-power" flag into the update
command, which should also make powerd permit the update.

BUG=b:193159451
TEST= tast run [user@]host[:port]
firmware.FwupdPowerdUpdateCheck.[case_name]

Change-Id: I593dfe8d3e702e620c1fc8ec016a28fc292e4f0b