camera: Add CCAUIDocumentScanning
This CL adds CCAUIDocumentScanning to test document scanning behavior in
CCA using File VCD to simulate the document in the scene.

However, since the "Document detection on preview" feature relies on CrOS
VCD, we cannot use File VCD to test that. Therefore, this CL only
targets on "Take photo and generate document file" part. For the preview
part, we plan to test it by CameraBox.

BUG=b:193196284
TEST=tast run [DUT] camera.CCAUIDocumentScanning

Change-Id: I2d48bf74442371af18b8b24520153f1752157703