Tast: Add partial gellerization test
This CL adds a Tast test for the gellerization flow. Currently it
launches and goes through the flow, stopping just before the last step.
(The account remains unsupervised at the end & can be used for
concurrent test runs.)  We're still researching how best to test the complete flow.

Note: This CL shouldn't be checked in before the private vars in
familylink.Gellerization.yaml are.

BUG=b:190654770
TEST=All familylink.* pass locally on VM.

Change-Id: I7829ccf457f471ff92798a982e6fe4e29ab28b51