tast-tests: Remove tablet_mode software dep from arcappcompat tests.
arcappcompat tests are intended to run only on compatible form factors
and used the tablet_mode software dep to request this. This doesn't work
for unibuild since it's satisfied on all models if any are convertible.
There isn't currently a hwdep that can handle this so in the meantime,
so depend on hwdep.TouchScreen() to at least filter out non-touch
devices, since all convertibles must have touchscreens, and remove
tablet_mode as a software dep.

BUG=b:189921066
TEST=None

Change-Id: I7d2421565ab33abcfa633a8b5fdf6997af0e6977