tast-tests: Migrating servo package from remote to common
Package servo contains the required utilities to communicate with servo
devices. At present, it lies inside chromiumos/tast/remote/. As per the
tast convention, packages inside remote should be used in remote
bundles but many local tests could also be benefitted from servo
(for example, tests that don't require a restart). That makes the common
directory an ideal place for servo package (following the convention).
So just for the sake of using a servo, tast tests don't have to be
written as remote from now on.

BUG=None
TEST=tast run --var=firmware.hibernate_time=20m --var=servo=$SERVO $DUT firmware.DeepSleep

Change-Id: Iac0f11d4039c4d22ffaf2dc10511bae71edb81ab