hardware: iioservice: remove warning when parsing trigger
When parsing /sys/bus/iio/devices, not only we encounter iio:deviceX,
we also list triggers (sysfs, hrtimer).
Add a new error type to return when probing a trigger when we expect a
device (in parseDevice) and vice-versa.
When we see a device but can not understand its internal, return a
"true" error as usual.

BUG=none
TEST=Check hardware.SensorIioservice and
hardware.SensorIioserviceHard pass without printing a warning message.

Change-Id: Ibe3a1a403cd7badae2503a12039851f02c4b2e6f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3055539
Tested-by: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Cheng-Hao Yang <chenghaoyang@chromium.org>