hardware: iioservice: remove warning when parsing trigger
When parsing /sys/bus/iio/devices, not only we encounter iio:deviceX,
we also list triggers (sysfs, hrtimer).
It is expected, ignore them when parsing the directory.
As a consequence, it is possible for parseSensor() to return an
empty sensor without error message.

BUG=none
TEST=Check hardware.SensorIioservice and
hardware.SensorIioserviceHard pass without printing a warning message.

Change-Id: Ibe3a1a403cd7badae2503a12039851f02c4b2e6f