network.DiagPassing: poll routines for expected result
Some routines test aspects of the network that are not completely in
control at the lab like DNS or https latency. This change instead polls
the routine to ensure the expected result can be achived. The goal of
these tests is not to test the network itself, but to test the framework
for running the routines.

BUG=chromium:1178896
TEST=tast run sarien-dut network.DiagPassing* (pass)

Change-Id: Ia02253210025b2da3ae0c9dddf6033337c44a52c