wificell: Add retry for connection to test APs
We're seeing several tests failed with "failed to connect to the router"
but there's no related logs on the router logs. In some logs, we see
that the later connections (e.g. after a few seconds) in the same task
can sometimes pass.

The cause of the flaky network is still unknown, but to mitigate the
impact of test coverage, let's add a retry here.

BUG=b:194536867
TEST=tast run $DUT wifi.SimpleConnect.80211a

Change-Id: I34c7af195de3b777844e369fe2f267f0177ab8bc