wificell: Add retry for connection to test APs
We're seeing several tests failed with "failed to connect to the router"
but there's no related logs on the router logs. In some logs, we see
that the later connections (e.g. after a few seconds) in the same task
can sometimes pass.

The cause of the flaky network is still unknown, but to mitigate the
impact of test coverage, let's add a retry for router connection. As to
pcap, as we can fall back to use router as pcap, let's not add retry
here for now to save the retry time on local setups without pcap.

BUG=b:194536867
TEST=tast run $DUT wifi.SimpleConnect.80211a

Change-Id: I34c7af195de3b777844e369fe2f267f0177ab8bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3057332
Tested-by: Yen-lin Lai <yenlinlai@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Commit-Queue: Yen-lin Lai <yenlinlai@chromium.org>