Move lacros binary extraction to Fixture time, not test run time.
Note that this doesn't add the Data field to the Fixtures, or update the
lacros.Setup() function. This change just does the non-mechanical part.

In a follow-up change, I will update the function signature of
lacros.Setup and add/remove Data fields to fixtures in the correct
places.

BUG=chromium:1233834
TEST=tast run <dut> ui.WindowCyclePerf

Change-Id: I7ae3a207602bb46a615ecb3cfaa0ba48700fd010
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3057349
Tested-by: Eliot Courtney <edcourtney@chromium.org>
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>