tast-tests: Add more logging to arc.TextToSpeech
This test is flaky on novato boards, but currently I haven't been able
to reproduce the failure on local workstation.
This change adds more logging in the test so that we can investigate
further.

The test is failing with "no such file or directory" error. So this
change adds logging around it.

Also, this tweaks file write options in the Android test app.

BUG=b:194309104
TEST=locally run novato vm, and `tast run localhost:<port> arc.TextToSpeech`

Change-Id: I734c4b21e6d984693ee7d73e94cc140afbb42f29
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3058598
Tested-by: Hiroki Sato <hirokisato@chromium.org>
Commit-Queue: Hiroki Sato <hirokisato@chromium.org>
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>
Reviewed-by: Oleh Lamzin <lamzin@google.com>