Allow ArcNotificationTest to generate high importance notifications.
This CL adds a checkbox, which when checked will generate a notification
with high importance. ARC notifications that with high importance are
shown as popup notifications.

This will be used in a tast test (which is to be written).

BUG=1233742
TEST=manual. Use test app to confirm that high important notifications
are sent

Change-Id: If8804d9d4c3beece49dff64737893d8bd4317b23
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3060134
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Hiroki Sato <hirokisato@chromium.org>
Tested-by: Sara Kato <sarakato@chromium.org>
Auto-Submit: Sara Kato <sarakato@chromium.org>
Commit-Queue: Sara Kato <sarakato@chromium.org>