faft: Disable CCD watchdog before flashing ap fw
When running firmware.ServoGBBFlags on leona, I noticed that the
ec console was read-only, and servod was broken. I.e. Running the test
once passed, but running it twice failed. Thost are classic symptoms of
the CCD watchdog detecting the EC being offline.

Changed ServoGBBFlags to remove the CCD watchdog before flashing
firmware.

Changed Servo.WatchdogRemove() to detect SuzyQ and remove the main
watchdog instead of ccd so that tests don't need that workaround.

BUG=b:185271203
TEST=tast run $DUT firmware.ServoGBBFlags

Change-Id: Ia6dc92b42344a25584d8050557d7b24ed05138d1