tast-tests: MeetCUJ assume 1st textfield to enter meeting code
Previously, the code assumes the meeting code field is the only
text field. Sometime, meet site would show a "Google Guide" widget
and breaks this assumption. Change the assumption to use the 1st
text field.

The assumption is necessary because the test has to run with
obfuscated meet frontend and the names/class names in a11y tree
are not human readable and keeps changing.

BUG=chromium:1234507
TEST=tast run $DUT ui.MeetCUU.*

Change-Id: I4335411bb9b0f2641815a8ef5556b72fbe9f185d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3061354
Tested-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Xiyuan Xia <xiyuan@chromium.org>