appcompat:Update Tast test for Microsoft Outlook
Microsoft Outlook is failing in launch. It is observed that the Next button present in the email address page and also in the password page aren’t visible on some devices and had to scroll down the page to make it visible. Added a code to fix it.

BUG=b:194107814
TEST=tast run <eve> arcappcompat.MicrosoftOutlook

Change-Id: I41d6311ba8156c34c16cf75692c8be15aee1fac5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3061356
Tested-by: Meenakshi Thiyagarajan <mthiyagarajan@google.com>
Reviewed-by: Renuga Nanjappan <rnanjappan@chromium.org>
Commit-Queue: Meenakshi Thiyagarajan <mthiyagarajan@google.com>