tast-test: Make setWindowStateR async
This CL changes setWindowStateR to async (just sending WMEvent and not
waiting for the window state change).
SetWindowState is expected to be async in the first place according to
the comment L514-515, and setWindowState*P* is async.
However, setWindowState*R* incorrectly run in a synchronous manner, and
it breaks arc.WindowState.tablet_vm.

BUG=b:179525662
TEST=arc.WindowState.tablet_vm

Cq-Depend: chromium:3061763
Change-Id: Id61e7a7c968da6dd8a4c370f99434d686c9c5bab