wifi.SimpleConnect: tast test for domain_suffix_match
This CL tests that, when domain_suffix_match is specified, the
wpa_supplicant will only connect to a wifi if one of the entries in
domain_suffix_match matches an entry in the dNSName element of
the SubjectAltName of the authentication server certificate.

To test against multiple valid entries in domain_suffix_match, this CL
updates the CertStore used for subjectAltName match testing by adding a
new DNS entry to the server certificates. Script used to generate the
certs: https://paste.googleplex.com/5073833313697792

It also adds additional unit tests, like checking that the expected
domain_suffix_match value used in the tast test is corresponding to a
dNSName entry in the test server certificate.

BUG=b:177651194,chromium:1203612
TEST=~/trunk/src/platform/tast/fast_build.sh -t \
     chromiumos/tast/common/wifi/security/wpaeap
TEST=~/trunk/src/platform/tast/fast_build.sh -t \
    chromiumos/tast/common/crypto/certificate
TEST=~/trunk/src/platform/tast/fast_build.sh -t \
    chromiumos/tast/common/wifi/security/tunneled1x
TEST=TODO: Run wifi.SimpleConnect tast tests in lab
    once implementation CLs land

Change-Id: I87d15ab9ff84f06e9b743f06d773df80d50e21c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3063299
Reviewed-by: Billy Zhao <billyzhao@chromium.org>
Reviewed-by: David Stevens <stevensd@chromium.org>
Tested-by: Andreea-Elena Costinas <acostinas@google.com>
Commit-Queue: Andreea-Elena Costinas <acostinas@google.com>