wifi.APSupportedRates: work around gopacket RTS incompatibilities
Our packet capture devices seem to be leaving off the checksum field at
times, and I believe this is tripping up gopacket -- it sees a
LayerTypeDot11, but not a LayerTypeDot11CtrlRTS, even when the Type
field is correct.

We can work around this by adding an extra filter to the Dot11 layer.

BUG=b:160281713
TEST=`wifi.APSupportedRates.*` on eve

Change-Id: If39da90f492ed886aa0d09d7980b370042284d74
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3063726
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Hsin-chen Chuang <chharry@google.com>
Reviewed-by: Trent Begin <tbegin@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>