arc: Change arc.AuthPerf.* boot_time calculation to use `ps` not `stat`
The current boot_time calculation gets the start time of the init
process using `stat -c %z /proc/1`. This is unreliable since filesystem
metadata is not immutable. Because of this, the boot_time is often
reported incorrectly, e.g. as ~60 s instead of the actual ~3 s.

Instead, get the init process start time using the `ps` command's
`stime` field.

BUG=b:188481433
TEST=Run arc.AuthPerf.managed on grunt-kernelnext. Without change,
     boot_time = 72685 ms. With change, boot_time = 2462 ms.

Change-Id: I7cde1aff669bc5b369f0b3a0f15aecb48ad0eb50