wificell: Capture packets during Roaming Continuity tests
It turned out that Roaming Continuity became so successful in
triggering various misbehaviors in the drivers that it is necessary
to collect pcap to discriminate various root causes.

As this test is not fully test fixture-compliant, the capture is not
triggered automatically by the fixture, but it must be started/stopped
manually.

BUG=b:175797315
TEST=wifi.RoamContPing.none

Change-Id: Ic9a43c09d3be7fe659eae96d6334efcc6df456f2