cdputil: Use "localhost" instead of "127.0.0.1"
Chrome devtools HTTP server rejects requests whose Host header is
neither a numerical address or "localhost" (crbug.com/813540).
For this reason, the cdp library (github.com/mafredri/cdp) resolves
a given hostname to a numerical address before sending HTTP requests
if it is not "localhost".
https://github.com/mafredri/cdp/blob/39a25b41c/devtool/devtool.go#L237

This resolution is performed even if a given hostname is a numerical
address, such as "127.0.0.1" we use today. And somehow we often see
"could not resolve IP for 127.0.0.1" error returned from this resolution
logic (b/194628311). I'm not exactly sure why this happens (possibly
infra DNS issues?)

Anyway, this patch tries to skip the resolution logic by changing the
hostname to "localhost".

BUG=b:194628311
TEST=tast run $DUT login.Chrome security.NetworkListeners*

Change-Id: I7b1ea147fea3d5f82f6c7a71dceef484d55bc891