video: support MemCheck on ARM platforms.
This CL teaches graphics/dri.go to use the DRM debug file
associated with the card "1" (previously it only looked for
the "0"), by looking in sequence for all such files.

The MemCheck test can then not be gated by "amd64" but rely
instead on hwdep.SupportsNV12Overlays() (that needs to be extended
a bit as per depend-CLs).

BUG=b:195755581
TEST=tast run <coachz ip> video.MemCheck.*_hw works !!

Change-Id: Iddd6c19251446366dc7bf47b11533beed9edda2f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3078894
Tested-by: Miguel Casas <mcasas@chromium.org>
Commit-Queue: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Po-Hsien Wang <pwang@chromium.org>