video: support MemCheck on ARM platforms.
This CL teaches graphics/dri.go to also use the framebuffer file
associated with the card "1" (previously it only looked for the "0").

The MemCheck test can then not be gated by "amd64" but rely
instead on hwdep.SupportsNV12Overlays() (that needs to be extended
a bit as per depend-CLs).

BUG=b:195755581
TEST=tast run <coachz ip> video.MemCheck.*_hw works !!

CQ-DEPEND=chromium:3078940,chromium:3078941

Change-Id: Iddd6c19251446366dc7bf47b11533beed9edda2f