camera: Clear contents in Camera folder but not the folder itself
By default, camera folder is not deletable by users. Therefore, we have
assumption that the camera folder won't be deleted during using CCA.

However, in https://crrev.com/c/3074384, we delete camera folder but not
relaunch CCA during sub tests for some tests, which break the assumption
mentioned above.

Therefore, this CL changes the behavior to only clear the contents in
Camera folder but not delete the folder itself.

BUG=b:195888225
TEST=tast run [DUT] camera.CCAUI{Resolutions,TakePicture}

Change-Id: I62176c675e4c45ba73b367346d77f6402e605214