kernel: Check EC device node existence
Adds a regression test for b/195345359. Note that unlike
firmware.ECVersion, this test does not depend on a servo, and is thus a
candidate for promotion to CQ inclusion.  It additionally does not
depend on ectool (good for increasing the likelihood that a failure is
related to a kernel change).

BUG=b:195345359,b:195967124
TEST=exercised the negative path (i.e. when /dev/cros_ec does exist) on
a local sona (so passed on a local sona).

Change-Id: I9035ee9046e2f9dd6cc1a89764b48dfa51d467e5