tast-tests: handle multiple entries in parsing sync-rtc-tlsdated-* files
Test platform.BootPerf is flaky with an error that it fails to parse the
sync-rtc-tlsdated-stop file because the file contains multiple
(typically 2) sync-rtc events. This is probably because some other test
stops/starts tlsdated and left sync-rtc-tlsdated entries under /tmp/.
Fix this flake by parsing the most recent entry (last line) in the
sync-rtc-tlsdated-* files.

BUG=b:196013597
TEST=restart tlsdated on DUT and tast run $IP_ADDR platform.BootPerf

Change-Id: I7624be80aeddfda19ef932244ae1a1211dfdc4ff