Tast: Direct set values for GAIA input fields using JS
Context in https://crbug.com/1222744. Previously we pulled up
the keyboard to type these values, which might contribute to
flakiness.

Only making this change for the GAIA fields. The parent
password (hosted by Unicorn) is still entered by keyboard because
we haven't seen flakiness on that page yet. (Using the keyboard also
mimics user behavior more closely, as opposed to The JS route, which
requires knowing more implementation details). If that page
turns out to be flaky as well, we can make the change later.

BUG=chromium:1222744
TEST=Affected familylink.* tests pass locally (ran each 5 times)
familylink.EducoexistenceInsession
familylink.NonEducoexistenceInSession

Not tested: arc.UnicornCannotAddNonEduAccount

Change-Id: I826f0cdd0a43528e1e4e501ce11bf390784ff7ca